Moved dto to content search domain
