+++ libs/api/domains/content-search/src/lib/dto/searcher.input.ts
