+++ libs/api/domains/content-search/src/lib/dto/webSearchAutocomplete.input.ts
@@ -1,3 +1,4 @@
+import { SearchIndexes } from '@island.is/content-search-indexer/types'
 import { Field, InputType, Int } from '@nestjs/graphql'
 import { IsEnum, IsInt, IsString, IsOptional } from 'class-validator'
 import { ContentLanguage } from '../enums/contentLanguage.enum'
