--- libs/content-search-toolkit/src/dto/index.ts
@@ -1,2 +1,0 @@
-export * from './searcher.input'
-export * from './webSearchAutocomplete.input'
