+++ libs/content-search-toolkit/src/index.ts
@@ -1,5 +1,3 @@
 export * from './services'
 export * from './types'
-export * from './dto'
-export * from './enums'
 export * from './elastic.module'
