+++ libs/content-search-toolkit/src/queries/dateAggregation.ts
@@ -1,25 +1,5 @@
-import { dateResolution, elasticTagField } from '../types'
+import { DateAggregationInput } from '../types'
 import { tagQuery } from './tagQuery'
-
-export interface DateAggregationInput {
-  types: string[]
-  field?: string
-  resolution?: dateResolution
-  order?: 'desc' | 'asc'
-  tags?: elasticTagField[]
-}
-
-type aggregationResult = {
-  key_as_string: string
-  key: number
-  doc_count: number
-}
-
-export interface DateAggregationResponse {
-  dates: {
-    buckets: aggregationResult[]
-  }
-}
 
 export const dateAggregationQuery = ({
   types = [],
