+++ libs/content-search-toolkit/src/queries/documentByMetaData.ts
@@ -1,27 +1,5 @@
-import { elasticTagField, sortableFields } from '../types'
+import { DocumentByMetaDataInput } from '../types'
 import { tagQuery } from './tagQuery'
-
-interface MetaPropsBase {
-  types?: string[]
-  tags?: elasticTagField[]
-  date?: {
-    from?: string
-    to?: string
-  }
-  sort?: sortableFields
-  page?: number
-  size?: number
-}
-
-interface RequiredTypes extends MetaPropsBase {
-  types: string[]
-}
-
-interface RequiredTags extends MetaPropsBase {
-  tags: elasticTagField[]
-}
-
-export type DocumentByMetaDataInput = RequiredTypes | RequiredTags
 
 // TODO: Make inclusivity optional
 // TODO: Scope tags to single document
