+++ libs/content-search-toolkit/src/types/autocomplete.ts
@@ -1,0 +1,10 @@
+export interface AutocompleteTermInput {
+  singleTerm: string
+  size: number
+}
+
+export interface AutocompleteTermResponse {
+  suggest: {
+    searchSuggester: [{ options: [{ text: string }] }]
+  }
+}
