+++ libs/content-search-toolkit/src/types/dateAggregation.ts
@@ -1,0 +1,21 @@
+import { dateResolution, elasticTagField } from '.';
+
+export interface DateAggregationInput {
+  types: string[]
+  field?: string
+  resolution?: dateResolution
+  order?: 'desc' | 'asc'
+  tags?: elasticTagField[]
+}
+
+type aggregationResult = {
+  key_as_string: string
+  key: number
+  doc_count: number
+}
+
+export interface DateAggregationResponse {
+  dates: {
+    buckets: aggregationResult[]
+  }
+}
