--- libs/content-search-toolkit/src/types/document.ts
@@ -1,31 +1,0 @@
-export interface Document {
-  _id: string | null
-  title: string
-  content: string
-  tag: [string]
-  date: string
-  date_updated: Date | null
-  category: string
-  other_categories: [string] | null
-  category_slug: string
-  category_description: string
-  subgroup: string
-  other_subgroups: [string] | null
-  containsApplicationForm: boolean
-  importance: number
-  group: string
-  group_slug: string
-  group_description: string
-  lang: string
-  url: string
-  slug: string
-  image: string
-  image_text: string
-  content_type: string
-  content_id: string
-  content_source: string
-  term_pool: string[]
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  content_blob: any
-  nextSyncToken?: string
-}
