+++ libs/content-search-toolkit/src/types/search.ts
@@ -1,0 +1,10 @@
+import { elasticTagField } from '.';
+
+export interface SearchInput {
+  queryString: string
+  size?: number
+  page?: number
+  types?: string[]
+  tags?: Omit<elasticTagField, 'value'>[]
+  countTag?: string
+}
