+++ libs/content-search-toolkit/src/types/tagAggregation.ts
@@ -1,0 +1,5 @@
+export interface TagAggregationInput {
+  documentTypes: string[]
+  tagType: string
+  size?: number
+}
